reorg accessibility bc
