--- subprojects/bounded_context/accessibility/domain/src/main/kotlin/club/staircrusher/accessibility/domain/service/PlaceAccessibilityService.kt
@@ -1,42 +1,0 @@
-package club.staircrusher.accessibility.domain.service
-
-import club.staircrusher.accessibility.domain.model.PlaceAccessibility
-import club.staircrusher.accessibility.domain.model.StairInfo
-import club.staircrusher.accessibility.domain.repository.PlaceAccessibilityRepository
-import club.staircrusher.stdlib.domain.DomainException
-import club.staircrusher.stdlib.domain.entity.EntityIdGenerator
-import club.staircrusher.stdlib.di.annotation.Component
-import java.time.Clock
-
-@Component
-class PlaceAccessibilityService(
-    private val clock: Clock,
-    private val placeAccessibilityRepository: PlaceAccessibilityRepository,
-) {
-    data class CreateParams(
-        val placeId: String,
-        val isFirstFloor: Boolean,
-        val stairInfo: StairInfo,
-        val hasSlope: Boolean,
-        val userId: String?,
-    )
-
-    fun create(params: CreateParams): PlaceAccessibility {
-        if (placeAccessibilityRepository.findByPlaceId(params.placeId) != null) {
-            throw DomainException("이미 접근성 정보가 등록된 장소입니다.")
-        }
-        val result = placeAccessibilityRepository.save(
-            PlaceAccessibility(
-                id = EntityIdGenerator.generateRandom(),
-                placeId = params.placeId,
-                isFirstFloor = params.isFirstFloor,
-                stairInfo = params.stairInfo,
-                hasSlope = params.hasSlope,
-                userId = params.userId,
-                createdAt = clock.instant(),
-            )
-        )
-
-        return result
-    }
-}
